+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * When for SimplyConnect pay buttons is set option "text", the plugin will not subscribe for total amount change event.
     * When the plugin can not find an Order by DMN data, do not try to create Order, but return 400 to the Cashier.
     * Implemented Auto-Void logic for DMNs.
+    * Trim merchant credentials when get them.
 ```
 
 # 3.1.0